% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\docType{methods}
\name{dbFetch,jdbcResult-method}
\alias{dbFetch,jdbcResult-method}
\title{Retrieve records from jdbc query}
\usage{
\S4method{dbFetch}{jdbcResult}(res, n = -1L, ...)
}
\description{
Retrieve records from jdbc query
}
