% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\docType{methods}
\name{dbSendQuery,jdbcConnection,character-method}
\alias{dbSendQuery,jdbcConnection,character-method}
\title{Send a query to jdbc.}
\usage{
\S4method{dbSendQuery}{jdbcConnection,character}(conn, statement, ...)
}
\description{
Send a query to jdbc.
}
