% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\docType{methods}
\name{dbSendStatement,jdbcConnection,character-method}
\alias{dbSendStatement,jdbcConnection,character-method}
\title{Send a statement}
\usage{
\S4method{dbSendStatement}{jdbcConnection,character}(conn, statement, ...)
}
\description{
Send a statement
}
